Update filenames and remove index files
