+++ samples/react-sample-app/src/ErrorBoundary.tsx
@@ -18,8 +18,7 @@
 
 import { AsgardeoAuthException } from "@asgardeo/auth-react";
 import React, { FunctionComponent, ReactElement } from "react";
-import { AuthenticationFailure } from "./pages/AuthenticationFailure";
-import { InvalidSystemTimePage } from "./pages/InvalidSystemTime";
+import { AuthenticationFailurePage, InvalidSystemTimePage } from "./pages";
 
 interface ErrorBoundaryProps {
     error: AsgardeoAuthException;
@@ -34,7 +33,7 @@ export const ErrorBoundary: FunctionComponent<ErrorBoundaryProps> = (
     if (error?.code === "SPA-CRYPTO-UTILS-VJ-IV01" || error?.message === "ERR_JWT_CLAIM_VALIDATION_FAILED nbf") {
         return <InvalidSystemTimePage/>
     } else if (error?.code === "SPA-MAIN_THREAD_CLIENT-SI-SE01") {
-        return <AuthenticationFailure/>
+        return <AuthenticationFailurePage/>
     }
 
     return children;
