+++ samples/react-sample-app/src/components/AgeVerificationDrawer.tsx
