+++ samples/react-sample-app/src/components/LoadingSpinner.tsx
