+++ samples/react-sample-app/src/components/ProtectedRoute.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Navigate, useLocation } from 'react-router-dom';
 import { useAuthContext } from "@asgardeo/auth-react";
-import { LoadingSpinner } from './loading-spinner';
+import { LoadingSpinner } from './LoadingSpinner';
 
 interface ProtectedRouteProps {
   children: React.ReactNode;
