+++ samples/react-sample-app/src/config-loader.ts
