+++ samples/react-sample-app/src/constants/idv-constants.ts
