+++ samples/react-sample-app/src/pages/InvalidSystemTimePage.tsx
@@ -18,7 +18,8 @@
 
 import React, { FunctionComponent, ReactElement } from "react";
 import { Button, Typography, Container, Box } from '@oxygen-ui/react';
-import { Footer, NavBar } from "../components";
+import { Footer } from "../components/Footer";
+import { NavBar } from "../components/NavBar";
 import { useNavigate } from "react-router-dom";
 
 /**
